backend/requirements.txt
Django==5.1.3
pymysql
djongo
python-dotenv
djangorestframework
djangorestframework-simplejwt